requirements.txt
Flask        1.0.2
requests     2.19.1
wtforms
flask-sqlalchemy
cymysql
flask-login
flask-mail